add mail test page
